teach-assist/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==1.3.5
Pillow==9.0.0
gocardless-pro==1.46.0 